I used eval typecast for getting integer as well as floating input and in html i use type=text

+++ Django/Calculator Using Django Form/calculator/app1/forms.py
@@ -2,6 +2,6 @@ from django import forms
 
 
 class CalculatorForm(forms.Form):
-    value1 = forms.IntegerField(label='Enter Value 1')
+    value1 = forms.FloatField(label='Enter Value 1')
     operator = forms.CharField(label='Enter Operator', max_length=10)
-    value2 = forms.IntegerField(label='Enter value 2')
+    value2 = forms.FloatField(label='Enter value 2')
